Not a finished, C -> Python returns mistake with int base_64 and 32

+++ loop.c
@@ -1,9 +1,9 @@
 #include <stdint.h>
 
-int loop(uint32_t numb)
+int loop(uint64_t numb)
 {
-    int res = 0;
-    int i = 0;
+    uint64_t res = 0;
+    uint64_t i = 0;
     for (i = 0; i < numb; i++)
     {
         res += i;
